Add Type property to Expression
I was stuck deciding how to solve the AST typing problem. I ended up just adding a nullable Type field to the abstract Expression class. While this is not "correct" it provides a simple solution that allows for me to continue work on the compiler while storing type information in the AST. This will probably be revisted in the future.